+++ chain-crypto/src/kes.rs
@@ -2,5 +2,5 @@ use crate::sign::VerificationAlgorithm;
 
 pub trait KeyEvolvingSignatureAlgorithm: VerificationAlgorithm {
     /// Get the period associated with this signature
-    fn get_period(sig: &Self::Signature) -> usize;
+    fn get_period(sig: &Self::Signature) -> u32;
 }
